Move pubkey utils to `tests/utils.rs`
.. as they are only used in tests.
