+++ src/utils.rs
@@ -1,10 +1,8 @@
-use bitcoin::secp256k1::PublicKey;
 use core::{fmt::Write, ops::Deref};
-use lightning::io;
 use lightning::sign::EntropySource;
 
 use crate::lsps0::msgs::RequestId;
-use crate::prelude::{String, Vec};
+use crate::prelude::String;
 
 /// Maximum transaction index that can be used in a `short_channel_id`.
 /// This value is based on the 3-bytes available for tx index.
@@ -54,53 +52,6 @@ pub fn hex_str(value: &[u8]) -> String {
 		write!(&mut res, "{:02x}", v).expect("Unable to write");
 	}
 	res
-}
-
-pub fn to_vec(hex: &str) -> Option<Vec<u8>> {
-	let mut out = Vec::with_capacity(hex.len() / 2);
-
-	let mut b = 0;
-	for (idx, c) in hex.as_bytes().iter().enumerate() {
-		b <<= 4;
-		match *c {
-			b'A'..=b'F' => b |= c - b'A' + 10,
-			b'a'..=b'f' => b |= c - b'a' + 10,
-			b'0'..=b'9' => b |= c - b'0',
-			_ => return None,
-		}
-		if (idx & 1) == 1 {
-			out.push(b);
-			b = 0;
-		}
-	}
-
-	Some(out)
-}
-
-pub fn to_compressed_pubkey(hex: &str) -> Option<PublicKey> {
-	if hex.len() != 33 * 2 {
-		return None;
-	}
-	let data = match to_vec(&hex[0..33 * 2]) {
-		Some(bytes) => bytes,
-		None => return None,
-	};
-	match PublicKey::from_slice(&data) {
-		Ok(pk) => Some(pk),
-		Err(_) => None,
-	}
-}
-
-pub fn parse_pubkey(pubkey_str: &str) -> Result<PublicKey, io::Error> {
-	let pubkey = to_compressed_pubkey(pubkey_str);
-	if pubkey.is_none() {
-		return Err(io::Error::new(
-			io::ErrorKind::Other,
-			"ERROR: unable to parse given pubkey for node",
-		));
-	}
-
-	Ok(pubkey.unwrap())
 }
 
 #[cfg(test)]
